feat: FileDelete applied on Menu, User

+++ src/main/java/shop/project/pathorderserver/_core/utils/FileUtil.java
@@ -14,7 +14,7 @@ import java.util.UUID;
 
 public class FileUtil {
 
-    public static String fileUpload(MultipartFile file) {
+    public static String uploadFile(MultipartFile file) {
         // UUID_파일 이름.확장자
         try {
             String newFilename = UUID.randomUUID() + "_" + file.getOriginalFilename();
@@ -27,13 +27,15 @@ public class FileUtil {
         }
     }
 
-    public static String base64Upload(String encodedData, String filename) {
+    public static String uploadBase64(String encodedData, String filename) {
         // UUID_파일 이름.확장자
         try {
             String mimeType = encodedData.substring(encodedData.indexOf(":") + 1, encodedData.indexOf(";"));
             MimeType type = MimeType.findByMimeType(mimeType);
+            if (type == null) { // 미확인 확장자 -> null return.
+                return null;
+            }
             String extension = type.getExtension();
-
             String newFilename = UUID.randomUUID() + "_" + filename + extension;
             Path newFilePath = Paths.get("./upload/" + newFilename);
 
@@ -46,7 +48,7 @@ public class FileUtil {
         }
     }
 
-    public static String base64JpgUpload(String encodedData, String filename) {
+    public static String uploadBase64Jpg(String encodedData, String filename) {
         // UUID_파일 이름.확장자 (* '.jpg'로 고정)
         String newFilename = UUID.randomUUID() + "_" + filename + MimeType.JPEG.getExtension();
         Path newFilePath = Paths.get("./upload/" + newFilename);
@@ -60,13 +62,14 @@ public class FileUtil {
         }
     }
 
-    public static void fileDelete(String filename) {
-        // TODO: default 파일 삭제 X 로직 추가
-        try {
-            Path filePath = Paths.get(getFilePath(filename));
-            Files.delete(filePath);
-        } catch (Exception e) {
-            throw new Exception500(e.getMessage());
+    public static void deleteFile(String filename) {
+        if (!DefaultFile.contains(filename)) { // 기존 파일이 아닌 경우에만 삭제
+            try {
+                Path filePath = Paths.get("." + getFilePath(filename));
+                Files.delete(filePath);
+            } catch (Exception e) {
+                throw new Exception500(e.getMessage());
+            }
         }
     }
